FC021: Check resource condition in LWRP
A change released in Chef 0.10.6 means that conditions may not work as expected
for resources redeclared with the same name. This rule checks providers for the
common case: a resource declared with a fixed name with a condition that is
calculated on the value of new_resource.

See this ticket for more background:
http://tickets.opscode.com/browse/CHEF-2812