FC042: Prefer include_recipe
This commit adds a Foodcritic rule to check for the use of
require_recipe in the recipe file and raise a deprecation warning.
All recipes should use include_recipe.
The require_recipe has been aliased to include_recipe since 2008
and will be removed as mentioned in [CHEF-3285](http://tickets.opscode.com/browse/CHEF-3285).