Update rules.rb
According to the man pages of wget: 
       Wget - The non-interactive network downloader.
This indeed is a downloader which should be constrained in a remote_file resource. 
However , 
According to the man pages of curl: 
curl is a tool to transfer data from or to a server, using one of the supported protocols
It nowhere explicitly is used as a downloader. 
It should be removed by definition from this rule. 
We could amend it by checking for curl  -o/--output <file>
This has been pointed out in : 
https://github.com/acrmp/foodcritic/issues/240
I would like to change this across food critic with a pull request. Please reply if you think that's acceptable