add warnings for use of recommends/suggests
These are unimplemented metadata keywords and their use is currently a
sort of no-op.  However, this can cause bugs with different tooling:

https://github.com/berkshelf/berkshelf/issues/1404

It also is likely that whatever the user thinks the semantics should be
eventually that some subset of users will guess incorrectly and get
surprising behavior if we ever do implement these keywords.  The proper
approach to documenting optional cookbooks is currently via text in the
README.md or a comment in the metadata.rb.

The correct course of action in the future is that an RFC should be
submitted to https://github.com/chef/chef-rfc to decide what the suggests
and recommends semantics are, and to actually implement those semantics.
Until then, users should avoid the use of these keywords (essentially
they are 'reserved for future use' -- which means "do not use them now").