Add self-dependency warning
Chef does not require cookbooks to depend on themselves in order to
depsolve.  In addition to this line being useless, it actually creates
problems with depsolving.  The self-dependency can create a circular
dependency where every version of a cookbook will satisfy the circular
dependency on itself which can create a huge mess.

Ideally we need to fix this elsewhere in the ecosystem as well, but
for now this will at least surface the issue to cookbook authors who
use foodcritic.