Fix FC058 false positives
There were multiple issues here. First, the test did not
correctly catch this because the wrong match specification
was being created due to line number not being specified.
Instead of defaulting to nil, it defaulted to 1, making
2 incorrect things look correct together.

Second, FC058 needed to check the length of the arrays
it was getting back was greater than 0, as [] defaults
to truthy.