new rule for Supermarket to check for open source license
Placed in only the supermarket category. Cookbooks shared on the public
Supermarket are expected to be shared under an open source license.
This rule uses the list of licenses from SPDX[1] that are identified as
OSI-approved. Ideally, this rule will be used to meet the license
quality metric[2] with a foodcritic run.

[1] https://spdx.org/licenses/
[2]
https://github.com/chef-cookbooks/cookbook-quality-metrics/blob/35bddc5af54eb8478b2307f7b6d231f03cf61ae9/quality-metrics/qm-003-license.md

Signed-off-by: Robb Kidd <robb@thekidds.org>