Remove Github Changelog Generator
A good changelog communicates what changed, why it was changes, and what the impact is. An autogenerated changelog communicates none of that and frankly github changelog generator has been nothing but trouble everywhere we’ve added it in Chef. It’s much easier to just write a changelog by hand where you convey the changes to the users and explain why it matters to them.

Signed-off-by: Tim Smith <tsmith@chef.io>