Properly detect notification actions that aren't symbols
Previously we did a query and whatever came back to .to_s.to_sym'd it. That meant even if nothing came back we ended up with :"" which is total junk. Now we properly find actions that are strings (which is sorta supported) and we return a nil if it's a variable since we have no way to properly handle that. This is MORE correct.

Signed-off-by: Tim Smith <tsmith@chef.io>